+++ sdk/storage/azure-storage-files-datalake/src/datalake_service_client.cpp
@@ -175,12 +175,12 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     Blobs::ListContainersSegmentOptions blobOptions;
     blobOptions.Context = options.Context;
     blobOptions.Prefix = options.Prefix;
-    blobOptions.ContinuationToken = options.Continuation;
+    blobOptions.ContinuationToken = options.ContinuationToken;
     blobOptions.MaxResults = options.MaxResults;
     auto result = m_blobServiceClient.ListBlobContainersSegment(blobOptions);
     auto response = ListFileSystemsSegmentResult();
-    response.Continuation
-        = result->ContinuationToken.empty() ? response.Continuation : result->ContinuationToken;
+    response.ContinuationToken
+        = result->ContinuationToken.empty() ? response.ContinuationToken : result->ContinuationToken;
     response.Filesystems = FileSystemsFromContainerItems(result->Items);
     return Azure::Core::Response<ListFileSystemsSegmentResult>(
         std::move(response), result.ExtractRawResponse());