+++ sdk/storage/azure-storage-files-datalake/CHANGELOG.md
@@ -6,16 +6,15 @@
 
 * Service version is now 2020-02-10.
 * Added support for SAS generation in DataLake service.
-* CreateFromConnectionString now accepts unencoded file, path and directory name.
 * Added support for `FileClient::ScheduleDeletion`.
 * Added support for `DirectoryClient::SetAccessControlRecursive`.
-* `ETag` and `LastModified` is now `std::string` instead of `Azure::Core::Nullable<std::string>` in `CreateDirectoryResult`, `CreateFileResult` and `CreatePathResult`.
 * `PathAppendDataResult` now returns `ContentMD5`, `ContentCrc64` and `IsServerEncrypted`.
 
 ### Breaking Changes
 
 * `CreateFromConnectionString` now accepts unencoded file, path and directory name.
 * `ETag` and `LastModified` is now `std::string` instead of `Azure::Core::Nullable<std::string>` in `CreateDirectoryResult`, `CreateFileResult` and `CreatePathResult`.
+* `Continuation` is renamed to `ContinuationToken` in options and returned result objects.
 
 ## 1.0.0-beta.2 (2020-09-09)